+++ modules/ui/admin/dashboard/ego.go
@@ -5,7 +5,7 @@ package dashboard
 
 import (
 	"fmt"
-	"github.com/infinitbyte/gopa/modules/ui/admin/common"
+	"github.com/infinitbyte/gopa/modules/ui/common"
 	"io"
 )